requirements.txt
apex==0.1
faiss==1.5.3
faiss_cpu==1.7.2
numpy==1.21.2
Pillow==9.2.0
scikit_learn==1.1.1
scipy==1.7.3
setuptools==58.0.4
tabulate==0.8.9
torch==1.8.2
torchvision==0.9.2
tqdm==4.62.3
yacs==0.1.8
